#78 jackson module was renamed to webflux-jackson
