+++ samples/coroutine-webapp/src/main/kotlin/org/springframework/fu/sample/coroutines/Application.kt
@@ -22,7 +22,7 @@ import org.springframework.context.event.ContextStartedEvent
 import org.springframework.fu.application
 import org.springframework.fu.module.data.mongodb.coroutine.coroutine
 import org.springframework.fu.module.data.mongodb.mongodb
-import org.springframework.fu.module.jackson.jackson
+import org.springframework.fu.module.webflux.jackson.jackson
 import org.springframework.fu.module.logging.*
 import org.springframework.fu.module.logging.LogLevel.*
 import org.springframework.fu.module.logging.logback.logback
