+++ samples/reactive-webapp/src/main/kotlin/org/springframework/fu/sample/reactive/Application.kt
@@ -20,7 +20,7 @@ import org.springframework.beans.factory.support.DefaultListableBeanFactory
 import org.springframework.context.event.ContextStartedEvent
 import org.springframework.fu.application
 import org.springframework.fu.module.data.mongodb.mongodb
-import org.springframework.fu.module.jackson.jackson
+import org.springframework.fu.module.webflux.jackson.jackson
 import org.springframework.fu.module.logging.*
 import org.springframework.fu.module.logging.LogLevel.*
 import org.springframework.fu.module.logging.logback.logback
