+++ settings.gradle.kts
@@ -9,12 +9,12 @@ include("bootstraps",
 		"modules:logging",
 		"modules:logging-logback",
 		"modules:dynamic-configuration",
-		"modules:jackson",
 		"modules:mongodb",
 		"modules:mongodb-coroutine",
 		"modules:mustache",
 		"modules:test",
 		"modules:webflux",
+		"modules:webflux-jackson",
 		"modules:webflux-coroutine",
 		"modules:webflux-netty",
 		"modules:webflux-jetty",
